+++ projects/ngx-draggable-dom/src/lib/ngx-draggable-dom.directive.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed } from "@angular/core/testing";
 
-import { NgxDraggableDomDirective } from "./ngx-draggable-dom.directive";
+import { NgxDraggableDomDirective } from "./directive/ngx-draggable-dom.directive";
 
 describe("NgxDraggableDomDirective", () => {
   beforeEach(() => TestBed.configureTestingModule({}));
