+++ projects/ngx-draggable-dom/src/public_api.ts
@@ -2,5 +2,5 @@
  * Public API Surface of ngx-draggable-dom
  */
 
-export * from "./lib/ngx-draggable-dom.directive";
+export * from "./lib/directive/ngx-draggable-dom.directive";
 export * from "./lib/ngx-draggable-dom.module";
